Refactor plugin config parsers to use disk util
We have a nice util to handle loading a file with a parser but the plugins
were not using it, this refactors them to use loadJson from the disk utils.

Test Plan: Ensure CLI is still able to load config files correctly
